ft(events): add events routes
