+++ src/routes/index.js
@@ -4,6 +4,7 @@ import image from "./image";
 import login from "./login";
 import message from "./message";
 import projects from "./projects";
+import events from "./event";
 
 const router = express.Router();
 
@@ -12,5 +13,6 @@ router.use("/images", image);
 router.use("/", login);
 router.use("/messages", message);
 router.use("/projects", projects);
+router.use("/events", events);
 
 export default router;
