+++ src/server.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 import express from "express";
 import logger from "morgan";
 import cors from "cors";
